Clean up documentation
